+++ CHANGELOG.md
@@ -17,6 +17,8 @@ This version is pinned strictly to `b2-sdk-python==1.17.3` for the same reason.
 * Add `--skip-hash-verification` parameter
 * Add `replication-status` command
 * Add `replication-delete` command
+* Add `replication-pause` command
+* Add `replication-unpause` command
 
 ### Changed
 * Minimum MacOS version from 10.15 to 11.0